add UnitOfWork and repository parttern

+++ ViewModel/Admin/Category/DeleteCategoryViewModel.cs
@@ -8,8 +8,8 @@ namespace ViewModel.Admin
 {
    public class DeleteCategoryViewModel
     {
-       public long Id { get; set; }
+       public int Id { get; set; }
        public string Name { get; set; }
-       public long ReplaceCategoryIdForProducts { get; set; }
+       public int ReplaceCategoryIdForProducts { get; set; }
     }
 }
